Added Cinemachine Player Tracking Camera
Imported the Cinemachine package, which allows for us to implement a camera that follows our player character. CM vcam1 is the component to edit, and you can edit how much dead space there is, which allows the camera to not move every time we move the player character, as well as turn on and off the guidelines that show up in the game view.